% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{applyMCMC}
\alias{applyMCMC}
\title{Apply a function to an mcmc or mcmc.list object, e.g. median}
\usage{
applyMCMC(data, fun = median, ...)
}
\arguments{
\item{data}{The object containing the MCMC sample - usually an object of class
"mcmc" or "mcmc.list"}

\item{fun}{The function to apply to the MCMC sample, by default median}

\item{...}{Extra parameters to be passed to the function}
}
\description{
Apply a function to an mcmc or mcmc.list object, e.g. median
}
