% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{checkESS}
\alias{checkESS}
\title{Check if any parameters in an mcmc chain have an ESS value < cutoff
Constant (unsampled) parameters are ignored.}
\usage{
checkESS(data, cutoff = 200, value = TRUE, ignored = c(), plot = FALSE, ...)
}
\arguments{
\item{data}{The object containing the MCMC sample - usually of class "mcmc" or "mcmc.list"}

\item{cutoff}{The cutoff ESS value (default = 200)}

\item{value}{If value == TRUE the parameters with ESS < cutoff and their ESS values are returned
If value == FALSE only the column indices of those parameters are returned}

\item{ignored}{Parameters in the chain that should be ignored}

\item{plot}{If plot == TRUE draw a barplot of the ESS values and mark parameters with ESS < cutoff}

\item{...}{Extra parameters to be passed to the plotESS() function}
}
\description{
Check if any parameters in an mcmc chain have an ESS value < cutoff
Constant (unsampled) parameters are ignored.
}
