% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{checkESSl}
\alias{checkESSl}
\title{Check that all parameters in an mcmc chain have an ESS value >= cutoff}
\usage{
checkESSl(data, cutoff = 200, ignored = c())
}
\description{
Check that all parameters in an mcmc chain have an ESS value >= cutoff
}
