% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{constantPars}
\alias{constantPars}
\title{Identify parameters in the mcmc chain that have not been sampled
i.e. parameters that have stayed constant across all chains}
\usage{
constantPars(data)
}
\arguments{
\item{data}{The object containing the MCMC sample - usually of class "mcmc" or "mcmc.list"}
}
\description{
If some parameters are constant on some chains but not on others
an error is thrown
}
