% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{constantParsMCMC}
\alias{constantParsMCMC}
\title{Like constantPars, but for a single mcmc object
(private helper function)}
\usage{
constantParsMCMC(data)
}
\description{
Like constantPars, but for a single mcmc object
(private helper function)
}
