% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{effectiveSize}
\alias{effectiveSize}
\title{Overwrites coda::effectiveSize
Avoids numerical underflow by scaling parameters so mean is 1 before calculating
the effecive size}
\usage{
effectiveSize(data, normalize = TRUE)
}
\arguments{
\item{data}{The object containing the MCMC sample - usually of class "mcmc" or "mcmc.list"}

\item{normalize}{If TRUE try to avoid numerical underflow, else just simply call
coda::effectiveSize}
}
\description{
Overwrites coda::effectiveSize
Avoids numerical underflow by scaling parameters so mean is 1 before calculating
the effecive size
}
