% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{getBranchingTimes}
\alias{getBranchingTimes}
\title{Get branching times of a tree}
\usage{
getBranchingTimes(tree, ...)
}
\description{
For an ultrametric tree this should be equivalent to the ape::coalescent.intervals
}
