% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{getCladeHeight}
\alias{getCladeHeight}
\title{Height of a clade in a phylogenetic tree}
\usage{
getCladeHeight(trees, tips, as.mcmc = TRUE)
}
\arguments{
\item{trees}{A single tree of class "phylo" or a set of trees of class "multiPhylo"}

\item{tips}{A vector of tips in the tree(s). Could either be numeric (tip numbers) or
a character vector (taxon names).}

\item{as.mcmc}{Return the distribution as an object of class "mcmc" (\code{\link[coda]{mcmc}}).
(only has an effect if \code{trees} is of class "multiPhylo").}
}
\value{
A numeric vector of heights or an "mcmc" object containing the heights.
}
\description{
Function that returns the height of the clade defined by the tips in \code{tips}.
If the input tree is a single phylogenetic tree of class "phylo" the height of
the clade is returned as a single number. If the input is a "multiPhylo" object
then the distribution of heights across all trees is returned. If \code{as.mcmc} is
true the distribution is returned as an "mcmc" object (\code{\link[coda]{mcmc}}).
}
\section{Warning}{

   If any of the tips are not present in any of the trees the function will crash.
}

\examples{

}
