% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{getCladeMonophyly}
\alias{getCladeMonophyly}
\title{Monophyly statistics of a clade in a list of trees}
\usage{
getCladeMonophyly(trees, tips, ...)
}
\arguments{
\item{trees}{A single tree of class "phylo" or a set of trees of class "multiPhylo"}

\item{tips}{A vector of tips in the tree(s). Could either be numeric (tip numbers) or
a character vector (taxon names).}
}
\value{
A floating point number between 0 and 1 giving the probability that the clade is
        monophyletic.
}
\description{
Function that returns the probsbility of the clade defined by the tips in \code{tips}
to be monophyletic in a list of trees. If the input tree is a single phylogenetic tree
of class "phylo" the function returns 0 or 1. If the input is a "multiPhylo" object
then the proportion of trees in which the clade is monophyletic is returned. This
function uses the \code{\link[ape]{is.monophyletic}} function.
}
\section{Warning}{

   If any of the tips are not present in any of the trees the function will crash.
}

\examples{

}
\seealso{
\code{\link[ape]{is.monophyletic}}
}
