% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{getHPDMedian}
\alias{getHPDMedian}
\title{Calculate the HPD interval and median and return in a single object}
\usage{
getHPDMedian(data, prob = 0.95, ...)
}
\arguments{
\item{data}{The object containing the MCMC sample - usually of class "mcmc" or "mcmc.list"}

\item{prob}{Target probability content for the HPD interval (see ?HPDinterval)}

\item{...}{Extra parameters to be passed to the median function}
}
\description{
Only works for mcmc and mcmc.list objects
}
