% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{getLineages}
\alias{getLineages}
\title{Internal function to get the number of lineages during an interval}
\usage{
getLineages(types)
}
\description{
Internal function to get the number of lineages during an interval
}
