% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{getMatrixHPD}
\alias{getMatrixHPD}
\title{Get HPD of a matrix of values e.g. a skyline}
\usage{
getMatrixHPD(data, margin = 2, dataframe = TRUE, ...)
}
\arguments{
\item{data}{The samples from the posterior. Can be in a data frame or in a matrix.
By default the function assumes that each row represents a sample from
the posterior and each column a parameter.}

\item{margin}{If the data are formatted with samples in columns set margin to 1.}

\item{dataframe}{If TRUE return output as a data frame, otherwise a matrix.}

\item{...}{Parameters to be passed to getHPD().}
}
\value{
3xn matrix or data frame where each column is c(lower, median, upper) for
        each parameter in turn.
}
\description{
Uses Chen and Shao algorithm as implemented in boa package.
}
