% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{getTreeIntervals}
\alias{getTreeIntervals}
\title{Return a table of the heights and types of all nodes in a tree}
\usage{
getTreeIntervals(tree, decreasing = FALSE)
}
\arguments{
\item{tree}{An object of class "phylo" from ape}
}
\description{
Internal nodes are labelled "coalescent" and leaves are lablled "sample"
}
\details{
Use depth first search to get the tree intervals

This can be used to get the vector of speciation/transmission and sampling times for a
phylogenetic tree.

The results of this function are equivalent to TreeSim::getx()

When sampling and coalescent events are at the same times the sampling events are considered first

The row names are the node numbers
}
