% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{getWeeks}
\alias{getWeeks}
\title{Generate a sequence of weeks}
\usage{
getWeeks(from, to, week_start = getOption("lubridate.week.start", 7),
  floorFrom = FALSE, ceilTo = FALSE)
}
\arguments{
\item{from, to}{Start and end of the date seq}

\item{week_start}{Specify the week start, default is Sunday (7)}

\item{floorFrom}{If TRUE the first element in the sequence is the start of
the first month before \code{from}, otherwise \code{from}.}

\item{ceilTo}{If TRUE the last element in the sequence is the start of
the first month after \code{to}, otherwise \code{to}.}
}
\description{
This function will return all the weeks between \code{from} and \code{to}.
The dates in the sequence are the first day of each week.
The first day of the week can be set using the \code{week_start} parameter.
}
