% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{pairwiseCorrelations}
\alias{pairwiseCorrelations}
\title{Calculate pairwise correlations between parameters}
\usage{
pairwiseCorrelations(data, ...)
}
\arguments{
\item{data}{Data frame or matrix of the posterior samples of parameters.
If data is a matrix, samples are rows and parameters are columns.}
}
\description{
Returns a matrix with the pairwise correlations between pairs of parameters.
This function is similar to \code{coda::corsscorr}, but unlike the
\code{coda} function other parameters can be passed to the correlation
function (so spearman or pearson correlation coefficients can be
calculated).
}
