% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{plotESS}
\alias{plotESS}
\title{Plot barplot of ESS values}
\usage{
plotESS(data, ess = NULL, cutoff = 200, title = "",
  col = mPal(dark$blue, 0.5), ...)
}
\arguments{
\item{data}{The object containing the MCMC sample - usually of class "mcmc" or "mcmc.list"}

\item{ess}{Output of running coda::effectiveSize(data) - this is optional and can be supplied
to save time}

\item{cutoff}{The cutoff ESS value (default = 200)}

\item{title}{Title of the plot (x-label)}

\item{col}{}
}
\description{
Plot barplot of ESS values
}
