% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{readLogfile}
\alias{readLogfile}
\title{Read in BEAST logfile}
\usage{
readLogfile(filename, burnin = 0.1, maxsamples = -1)
}
\arguments{
\item{filename}{The logfile.}

\item{burnin}{Discard this proportion of samples at the start of the chain.}

\item{maxamples}{If > 0 stop after reading in this many lines
(this option is only for testing and should generally not be used).}
}
\description{
Read in BEAST logfile
}
