% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{sampledPars}
\alias{sampledPars}
\title{The opposite to constantPars, that is returns all parameters
that have been sampled}
\usage{
sampledPars(data)
}
\arguments{
\item{data}{The object containing the MCMC sample - usually of class "mcmc" or "mcmc.list"}
}
\description{
The opposite to constantPars, that is returns all parameters
that have been sampled
}
