% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{treeDFS}
\alias{treeDFS}
\title{Internal function for depth first traversal of a tree in order to get
the node heights}
\usage{
treeDFS(i, t, treetable)
}
\description{
Internal function for depth first traversal of a tree in order to get
the node heights
}
