% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{treeDFSSlow}
\alias{treeDFSSlow}
\title{Internal function for depth first traversal of a tree in order to get
the node heights (for treeIntervalsSlow)}
\usage{
treeDFSSlow(i, t, treetable)
}
\description{
Internal function for depth first traversal of a tree in order to get
the node heights (for treeIntervalsSlow)
}
