% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{writeLog}
\alias{writeLog}
\title{Save BEAST/Tracer log file}
\usage{
writeLog(logfile, filename, renumber = FALSE, header = TRUE, model = "")
}
\arguments{
\item{renumber}{Renumber states from 1 to nrow(logfile)}

\item{header}{Add a header with summary statistics to the file}

\item{model}{Add the model description from a logfile written in BEAST (not implemented)}
}
\description{
What happens if it's a mcmc.list or a list of data frames?
What if input is just a data frame?
}
