% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{getLogFileSubset}
\alias{getLogFileSubset}
\title{Extract all matching parameters from the logfile}
\usage{
getLogFileSubset(logfile, par)
}
\arguments{
\item{par}{The string to match with the start of the parameters to extract.}
}
\value{
A number of parameters from the logfile that match "^<par>"
}
\description{
if par="R0" extract (R0s.1 R0s.2 R0s.3 etc.)
}
