% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{getMonths}
\alias{getMonths}
\title{Generate a sequence of months}
\usage{
getMonths(from, to, floorFrom = TRUE, ceilTo = TRUE)
}
\arguments{
\item{from, to}{Start and end of the date seq}

\item{floorFrom}{If TRUE the first element in the sequence is the start of
the first month before \code{from}, otherwise \code{from}.}

\item{ceilTo}{If TRUE the last element in the sequence is the start of
the first month after \code{to}, otherwise \code{to}.}
}
\description{
This function will return all the months between \code{from} and \code{to}.
The dates in the sequence are the first day of each month.
}
