% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{readSingleLog}
\alias{readSingleLog}
\title{Read a single BEAST log file}
\usage{
readSingleLog(
  filename,
  burnin = 0.1,
  maxsamples = -1,
  as.mcmc = TRUE,
  burninAsSamples = FALSE
)
}
\arguments{
\item{filename}{The name of the log file to read.}

\item{burnin}{Discard this proportion of samples at the start of the chain
(if `burninAsSamples` == FALSE). Otherwise discard this many samples
at the start of the chain.}

\item{maxsamples}{If > 0 stop after reading in this many lines
(this option is only for testing and should generally not be used).}

\item{as.mcmc}{If FALSE then return an object of class "data.frame", else
return an "mcmc" object}

\item{burninAsSamples}{if TRUE burnin is given as the number of samples,
if FALSE burnin is a proportion (0 <= burnin < 1) of samples.
(default = FALSE).}
}
\value{
An "mcmc" object (\code{\link[coda]{mcmc}}) or data frame
        (\code{\link{data.frame}}) object containing all of the parameters
        in the MCMC chain, with the burn-in discarded.
}
\description{
Read a single BEAST log file and return as a coda "mcmc" object.
}
\examples{

}
\seealso{
\code{\link{readLog}}
}
