% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{readTreeLog}
\alias{readTreeLog}
\title{Read in BEAST trees file}
\usage{
readTreeLog(filename, burnin = 0.1)
}
\arguments{
\item{filename}{The trees file.}

\item{burnin}{Discard this proportion of samples at the start of the chain.}
}
\description{
The file can contain a single tree or a set of posterior trees. The
file must be in NEXUS format and must be terminated by END;
}
\section{Warning}{

  The function will attempt to read the whole trees file. If the
  file is large this may take a long time or run out of memory.
}

